filter university, speciality, program

+++ diploma/database/migrations/2022_05_04_141227_create_speciality_program_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('speciality_program', function (Blueprint $table) {
             $table->id();
-            $table->integer('program_id');
-            $table->integer('speciality_id');
+            $table->unsignedBigInteger('program_id');
+            $table->unsignedBigInteger('speciality_id');
         });
     }
